1/service_worker/requirements.txt
SQLAlchemy==2.0.29
psycopg2-binary==2.9.11
boto3==1.34.69
pika==1.3.2
requests==2.32.3
python-dotenv
Pillow>=10.0.0